NumberUnit Datums are always backed by a Double now.
svn path=/dasCore/trunk/; revision=966

+++ dasCore/src/edu/uiowa/physics/pw/das/datum/NumberUnits.java
@@ -46,11 +46,11 @@ public class NumberUnits extends Units {
     }
     
     public Datum createDatum( int value ) {
-        return new Datum.Double( new Double(value), this );
+        return new Datum.Double( new Long(value), this );
     }
     
     public Datum createDatum( long value ) {
-        return new Datum.Double( new Double(value), this );
+        return new Datum.Double( new Long(value), this );
     }
     
     public Datum createDatum( Number value ) {
